UI for desired rung offset; dynamized inner trim
You can now specify in pixels the desired width of a gap or overlap between the rungs of the spiral. It calculates a growth increment to match the desired rung gap.  negative values are read as overlap. the central circular trim  at the end now also derives from this gap.